Only show auth0 part of the profile in the output for auth0 output Change group handling for backward compat and clarity on group quality.
The idea being this is:

"groups" are self managed groups in either LDAP or Mozillians.org and
let RPs decide how to best manage access themselves. For ex. discourse
can create a Mozillians.org group "IAM" and both manage and use it to
restrict access to certain discussions to accounts in that group.

"authoritative_groups" are strongly managed groups that are emitted by
IAM directly. This includes the "moco", "mofo" groups (ie Staff members)
that is sourced from workday/servicenow, the "nda" group which is well
managed and sourced from mozillians.org  and all the auth0 service
groups which are created for each RP.

Authentication is done via 2 layers:
- Auth0: allows user to login if the user is in any of the authoritative
groups - if none is required, just allow login. When logging the user
in, the "lastUsed" attribute is also updated to reflect the last time
the user used that group.
If "lastUsed" is older than 3month, access is denied and the user must
re-request access

- RP: the rp can use any of the "groups" or "authoritative_groups" at
it's entire discretion for additional access management purposes.